automessage-api/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
distro==1.9.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
groq==0.9.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
kombu==5.4.2
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==2.0.1
openai==1.40.3
optree==0.12.1
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
pillow==10.4.0
prompt_toolkit==3.0.48
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.10.0
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-Levenshtein==0.25.1
python-socketio==5.11.4
pytz==2024.1
pywin32-ctypes==0.2.3
rapidfuzz==3.9.6
redis==5.2.0
requests==2.32.3
rich==13.7.1
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==73.0.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
